challenge 1/backend/requirements.txt


Flask
Flask-CORS
openai>=1.0.0
sentence-transformers
scikit-learn
scipy
transformers
python-dotenv
numpy